requirements.txt
yt-dlp>=2024.1.0
faster-whisper>=1.0.0
tushare>=1.4.0
python-dotenv>=1.0.0
